requirements.txt
discord.py
PyNaCl==1.3.0
dnspython==1.16.0
pytz==2021.1
youtube-dl==2020.12.2
DiscordUtils==1.3.4

